feat: Refactor chat handling by removing unused factories and enhancing message processing
